Enum linera_alloy_json_abi::AbiItem

pub enum AbiItem<'a> {
    Constructor(Cow<'a, Constructor>),
    Fallback(Cow<'a, Fallback>),
    Receive(Cow<'a, Receive>),
    Function(Cow<'a, Function>),
    Event(Cow<'a, Event>),
    Error(Cow<'a, Error>),
}

A JSON ABI item.

Variants

Constructor(Cow<'a, Constructor>)

A JSON ABI Constructor.

Fallback(Cow<'a, Fallback>)

A JSON ABI Fallback.

Receive(Cow<'a, Receive>)

A JSON ABI Receive.

Function(Cow<'a, Function>)

A JSON ABI Function.

Event(Cow<'a, Event>)

A JSON ABI Event.

Error(Cow<'a, Error>)

A JSON ABI Error.

Implementations

impl AbiItem<'_>

pub fn parse(input: &str) -> Result<Self>

Parses a single Human-Readable ABI string into an ABI item.

Examples

assert_eq!(
    AbiItem::parse("function foo(bool bar)"),
    Ok(AbiItem::from(Function::parse("foo(bool bar)").unwrap()).into()),
);

pub const fn debug_name(&self) -> &'static str

Returns the debug name of the item.

pub fn name(&self) -> Option<&String>

Returns an immutable reference to the name of the item.

pub fn name_mut(&mut self) -> Option<&mut String>

Returns a mutable reference to the name of the item.

Clones the item if it is not already owned.

pub fn state_mutability(&self) -> Option<StateMutability>

Returns the state mutability of the item.

pub fn state_mutability_mut(&mut self) -> Option<&mut StateMutability>

Returns a mutable reference to the state mutability of the item.

Clones the item if it is not already owned.

pub fn inputs(&self) -> Option<&Vec<Param>>

Returns an immutable reference to the inputs of the item.

Use event_inputs for events instead.

pub fn inputs_mut(&mut self) -> Option<&mut Vec<Param>>

Returns a mutable reference to the inputs of the item.

Clones the item if it is not already owned.

Use event_inputs for events instead.

pub fn event_inputs(&self) -> Option<&Vec<EventParam>>

Returns an immutable reference to the event inputs of the item.

Use inputs for other items instead.

pub fn event_inputs_mut(&mut self) -> Option<&mut Vec<EventParam>>

Returns a mutable reference to the event inputs of the item.

Clones the item if it is not already owned.

Use inputs for other items instead.

pub fn outputs(&self) -> Option<&Vec<Param>>

Returns an immutable reference to the outputs of the item.

pub fn outputs_mut(&mut self) -> Option<&mut Vec<Param>>

Returns an immutable reference to the outputs of the item.